requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
pymongo==4.7.2
python-dotenv==1.0.1
sendgrid==6.11.0
